Formatting urlpart rules modified
If a user manualy typed a url part this one will be used. If the urlpart
is empty then the title will be used as urlpart. In any case the
StrinHelper will modify the urpart, even is the user already gave an
urlpart, just to be sure it is an valid url.